+++ common/src/main/java/com/tristankechlo/livingthings/config/entity/OwlConfig.java
@@ -2,13 +2,10 @@ package com.tristankechlo.livingthings.config.entity;
 
 import com.tristankechlo.livingthings.config.util.EntityConfig;
 import com.tristankechlo.livingthings.config.util.SpawnData;
-import com.tristankechlo.livingthings.config.values.IngredientValue;
 import com.tristankechlo.livingthings.config.values.ListValue;
 import com.tristankechlo.livingthings.config.values.NumberValue.DoubleValue;
 import com.tristankechlo.livingthings.config.values.NumberValue.IntegerValue;
 import net.minecraft.resources.ResourceKey;
-import net.minecraft.world.item.Items;
-import net.minecraft.world.item.crafting.Ingredient;
 import net.minecraft.world.level.biome.Biomes;
 
 import java.util.List;
@@ -21,8 +18,6 @@ public final class OwlConfig extends EntityConfig {
     public final DoubleValue movementSpeed = new DoubleValue("movementSpeed", 0.25D, MIN_SPEED, MAX_SPEED);
     public final DoubleValue flyingSpeed = new DoubleValue("flyingSpeed", 0.5D, MIN_SPEED, MAX_SPEED);
     public final IntegerValue maxSpawnedInChunk = new IntegerValue("maxSpawnedInChunk", 6, 1, 15);
-    public final IngredientValue temptationItems = new IngredientValue("temptationItems", Items.MELON_SEEDS, Items.PUMPKIN_SEEDS, Items.BEETROOT_SEEDS);
-    public final IngredientValue tamingItems = new IngredientValue("tamingItems", Items.WHEAT_SEEDS);
     public final ListValue<SpawnData> spawnBiomes = new ListValue<>("spawnBiomes", createDefaultSpawns(), SpawnData::serialize, SpawnData::deserialize);
 
     public final IntegerValue colorWhiteWeight = new IntegerValue("colorWhiteWeight", 33, 0, Integer.MAX_VALUE);
@@ -31,7 +26,7 @@ public final class OwlConfig extends EntityConfig {
 
     private OwlConfig() {
         super("owl");
-        this.registerConfigValues(health, movementSpeed, flyingSpeed, maxSpawnedInChunk, temptationItems, tamingItems, spawnBiomes);
+        this.registerConfigValues(health, movementSpeed, flyingSpeed, maxSpawnedInChunk, spawnBiomes);
         this.registerForCategory("colorWeights", colorWhiteWeight, colorBrownWeight, colorBlackWeight);
     }
 
@@ -53,14 +48,6 @@ public final class OwlConfig extends EntityConfig {
 
     public static int maxSpawnedInChunk() {
         return INSTANCE.maxSpawnedInChunk.get();
-    }
-
-    public static Ingredient temptationItems() {
-        return INSTANCE.temptationItems.get();
-    }
-
-    public static Ingredient tamingItems() {
-        return INSTANCE.tamingItems.get();
     }
 
     private static List<SpawnData> createDefaultSpawns() {