+++ common/src/main/java/com/tristankechlo/livingthings/entity/KoalaEntity.java
@@ -3,6 +3,7 @@ package com.tristankechlo.livingthings.entity;
 import com.tristankechlo.livingthings.config.entity.KoalaConfig;
 import com.tristankechlo.livingthings.init.ModEntityTypes;
 import com.tristankechlo.livingthings.util.ILexiconEntry;
+import com.tristankechlo.livingthings.util.Ingredients;
 import com.tristankechlo.livingthings.util.LexiconEntries;
 import com.tristankechlo.livingthings.util.LivingThingsTags;
 import net.minecraft.core.BlockPos;
@@ -48,7 +49,7 @@ public class KoalaEntity extends Animal implements ILexiconEntry {
         this.goalSelector.addGoal(0, new FloatGoal(this));
         this.goalSelector.addGoal(1, new PanicGoal(this, 1.25D));
         this.goalSelector.addGoal(2, new BreedGoal(this, 1.0D));
-        this.goalSelector.addGoal(3, new TemptGoal(this, 1.1D, KoalaConfig.temptationItems(), false));
+        this.goalSelector.addGoal(3, new TemptGoal(this, 1.1D, Ingredients.KOALA_FOOD, false));
         this.goalSelector.addGoal(4, new FollowParentGoal(this, 1.1D));
         this.goalSelector.addGoal(5, new WaterAvoidingRandomStrollGoal(this, 1.0D));
         this.goalSelector.addGoal(6, new LookAtPlayerGoal(this, Player.class, 6.0F));
@@ -80,7 +81,7 @@ public class KoalaEntity extends Animal implements ILexiconEntry {
 
     @Override
     public boolean isFood(ItemStack stack) {
-        return KoalaConfig.temptationItems().test(stack);
+        return stack.is(LivingThingsTags.KOALA_FOOD);
     }
 
     @Override