+++ common/src/main/java/com/tristankechlo/livingthings/entity/ShroomieEntity.java
@@ -5,6 +5,7 @@ import com.tristankechlo.livingthings.entity.ai.ShroomiePlantMushroomGoal;
 import com.tristankechlo.livingthings.entity.misc.IMobVariants;
 import com.tristankechlo.livingthings.init.ModEntityTypes;
 import com.tristankechlo.livingthings.util.ILexiconEntry;
+import com.tristankechlo.livingthings.util.Ingredients;
 import com.tristankechlo.livingthings.util.LexiconEntries;
 import com.tristankechlo.livingthings.util.LivingThingsTags;
 import net.minecraft.core.BlockPos;
@@ -69,7 +70,7 @@ public class ShroomieEntity extends Animal implements IMobVariants, ILexiconEntr
         this.goalSelector.addGoal(0, new FloatGoal(this));
         this.goalSelector.addGoal(1, new PanicGoal(this, 1.25D));
         this.goalSelector.addGoal(2, new BreedGoal(this, 1.0D));
-        this.goalSelector.addGoal(3, new TemptGoal(this, 1.1D, ShroomieConfig.temptationItems(), false));
+        this.goalSelector.addGoal(3, new TemptGoal(this, 1.1D, Ingredients.SHROOMIE_FOOD, false));
         this.goalSelector.addGoal(4, new FollowParentGoal(this, 1.1D));
         this.goalSelector.addGoal(5, new ShroomiePlantMushroomGoal(this));
         this.goalSelector.addGoal(7, new WaterAvoidingRandomStrollGoal(this, 1.0D));
@@ -101,7 +102,7 @@ public class ShroomieEntity extends Animal implements IMobVariants, ILexiconEntr
 
     @Override
     public boolean isFood(ItemStack stack) {
-        return ShroomieConfig.temptationItems().test(stack);
+        return stack.is(LivingThingsTags.SHROOMIE_FOOD);
     }
 
     @Override