+++ fabric/src/main/java/com/tristankechlo/livingthings/platform/FabricPlatformHelper.java
@@ -1,5 +1,6 @@
 package com.tristankechlo.livingthings.platform;
 
+import com.google.auto.service.AutoService;
 import com.tristankechlo.livingthings.entity.SeahorseEntity;
 import com.tristankechlo.livingthings.init.ModItems;
 import net.fabricmc.fabric.api.itemgroup.v1.FabricItemGroup;
@@ -9,7 +10,6 @@ import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.sounds.SoundEvent;
-import net.minecraft.tags.TagKey;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.Mob;
 import net.minecraft.world.item.CreativeModeTab;
@@ -21,16 +21,12 @@ import net.minecraft.world.level.material.Fluid;
 import java.nio.file.Path;
 import java.util.function.Supplier;
 
+@AutoService(IPlatformHelper.class)
 public final class FabricPlatformHelper implements IPlatformHelper {
 
     @Override
     public String getPlatformName() {
         return "Fabric";
-    }
-
-    @Override
-    public TagKey<Item> getBananaTag() {
-        return TagKey.create(Registries.ITEM, new ResourceLocation("c", "fruits/bananas"));
     }
 
     @Override